backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-magic
python-dotenv
python-multipart   
PyMuPDF